requirements.txt
# Deep Learning & Numerical Processing
tensorflow>=2.10.0
numpy
scikit-learn

# Image Processing
Pillow